Reduce use of local variables in graph editor export
The local variables can clash with muparserx builtins
because they come from a diverse set of blocks not
knowing that their parameter names might be variable constants.

The GUI will usually avoid creating local variable evaluators
for parsing parameter expressions if the expression is simply
a single parameter.

This change reduces the use of local variables when exporting the
topology so that the same basic rules will apply in the exported design.